+++ pkg/planner/core/hint_utils.go
@@ -1,0 +1,416 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package core
+
+import (
+	"bytes"
+	"fmt"
+	"strings"
+
+	"github.com/pingcap/tidb/pkg/kv"
+	"github.com/pingcap/tidb/pkg/parser/ast"
+	"github.com/pingcap/tidb/pkg/parser/model"
+	"github.com/pingcap/tidb/pkg/sessionctx"
+	utilhint "github.com/pingcap/tidb/pkg/util/hint"
+)
+
+// GenHintsFromFlatPlan generates hints from a FlatPhysicalPlan.
+func GenHintsFromFlatPlan(flat *FlatPhysicalPlan) []*ast.TableOptimizerHint {
+	if len(flat.Main) == 0 {
+		return nil
+	}
+	nodeTp := utilhint.TypeSelect
+	switch flat.Main[0].Origin.(type) {
+	case *Update:
+		nodeTp = utilhint.TypeUpdate
+	case *Delete:
+		nodeTp = utilhint.TypeDelete
+	}
+	var hints []*ast.TableOptimizerHint
+	selectPlan, _ := flat.Main.GetSelectPlan()
+	if len(selectPlan) == 0 || !selectPlan[0].IsPhysicalPlan {
+		return nil
+	}
+	for _, op := range selectPlan {
+		p := op.Origin.(PhysicalPlan)
+		hints = genHintsFromSingle(p, nodeTp, op.StoreType, hints)
+	}
+	for _, cte := range flat.CTEs {
+		for i, op := range cte {
+			if i == 0 || !op.IsRoot {
+				continue
+			}
+			p := op.Origin.(PhysicalPlan)
+			hints = genHintsFromSingle(p, nodeTp, op.StoreType, hints)
+		}
+	}
+	return removeDuplicatedHints(hints)
+}
+
+// GenHintsFromPhysicalPlan generates hints from physical plan.
+func GenHintsFromPhysicalPlan(p Plan) []*ast.TableOptimizerHint {
+	flat := FlattenPhysicalPlan(p, false)
+	return GenHintsFromFlatPlan(flat)
+}
+
+func getTableName(tblName model.CIStr, asName *model.CIStr) model.CIStr {
+	if asName != nil && asName.L != "" {
+		return *asName
+	}
+	return tblName
+}
+
+func extractTableAsName(p PhysicalPlan) (*model.CIStr, *model.CIStr) {
+	if len(p.Children()) > 1 {
+		return nil, nil
+	}
+	switch x := p.(type) {
+	case *PhysicalTableReader:
+		ts := x.TablePlans[0].(*PhysicalTableScan)
+		if ts.TableAsName.L != "" {
+			return &ts.DBName, ts.TableAsName
+		}
+		return &ts.DBName, &ts.Table.Name
+	case *PhysicalIndexReader:
+		is := x.IndexPlans[0].(*PhysicalIndexScan)
+		if is.TableAsName.L != "" {
+			return &is.DBName, is.TableAsName
+		}
+		return &is.DBName, &is.Table.Name
+	case *PhysicalIndexLookUpReader:
+		is := x.IndexPlans[0].(*PhysicalIndexScan)
+		if is.TableAsName.L != "" {
+			return &is.DBName, is.TableAsName
+		}
+		return &is.DBName, &is.Table.Name
+	case *PhysicalSort, *PhysicalSelection, *PhysicalUnionScan, *PhysicalProjection:
+		return extractTableAsName(p.Children()[0])
+	}
+	return nil, nil
+}
+
+func getJoinHints(sctx sessionctx.Context, joinType string, parentOffset int, nodeType utilhint.NodeType, children ...PhysicalPlan) (res []*ast.TableOptimizerHint) {
+	if parentOffset == -1 {
+		return res
+	}
+	for _, child := range children {
+		blockOffset := child.SelectBlockOffset()
+		if blockOffset == -1 {
+			continue
+		}
+		var dbName, tableName *model.CIStr
+		if blockOffset != parentOffset {
+			var blockAsNames []ast.HintTable
+			if p := sctx.GetSessionVars().PlannerSelectBlockAsName.Load(); p != nil {
+				blockAsNames = *p
+			}
+			if blockOffset >= len(blockAsNames) {
+				continue
+			}
+			hintTable := blockAsNames[blockOffset]
+			// For sub-queries like `(select * from t) t1`, t1 should belong to its surrounding select block.
+			dbName, tableName, blockOffset = &hintTable.DBName, &hintTable.TableName, parentOffset
+		} else {
+			dbName, tableName = extractTableAsName(child)
+		}
+		if tableName == nil || tableName.L == "" {
+			continue
+		}
+		qbName, err := utilhint.GenerateQBName(nodeType, blockOffset)
+		if err != nil {
+			continue
+		}
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(joinType),
+			Tables:   []ast.HintTable{{DBName: *dbName, TableName: *tableName}},
+		})
+		break
+	}
+	return res
+}
+
+func genHintsFromSingle(p PhysicalPlan, nodeType utilhint.NodeType, storeType kv.StoreType, res []*ast.TableOptimizerHint) []*ast.TableOptimizerHint {
+	qbName, err := utilhint.GenerateQBName(nodeType, p.SelectBlockOffset())
+	if err != nil {
+		return res
+	}
+	switch pp := p.(type) {
+	case *PhysicalLimit, *PhysicalTopN:
+		if storeType == kv.TiKV {
+			res = append(res, &ast.TableOptimizerHint{
+				QBName:   qbName,
+				HintName: model.NewCIStr(HintLimitToCop),
+			})
+		}
+	case *PhysicalTableReader:
+		tbl, ok := pp.TablePlans[0].(*PhysicalTableScan)
+		if !ok {
+			return res
+		}
+		if tbl.StoreType == kv.TiFlash {
+			res = append(res, &ast.TableOptimizerHint{
+				QBName:   qbName,
+				HintName: model.NewCIStr(HintReadFromStorage),
+				HintData: model.NewCIStr(kv.TiFlash.Name()),
+				Tables:   []ast.HintTable{{DBName: tbl.DBName, TableName: getTableName(tbl.Table.Name, tbl.TableAsName)}},
+			})
+		} else {
+			res = append(res, &ast.TableOptimizerHint{
+				QBName:   qbName,
+				HintName: model.NewCIStr(HintUseIndex),
+				Tables:   []ast.HintTable{{DBName: tbl.DBName, TableName: getTableName(tbl.Table.Name, tbl.TableAsName)}},
+			})
+			if tbl.Table.PKIsHandle || tbl.Table.IsCommonHandle { // it's a primary key
+				orderHint := HintOrderIndex
+				if !tbl.KeepOrder {
+					orderHint = HintNoOrderIndex
+				}
+				res = append(res, &ast.TableOptimizerHint{
+					QBName:   qbName,
+					HintName: model.NewCIStr(orderHint),
+					Tables:   []ast.HintTable{{DBName: tbl.DBName, TableName: getTableName(tbl.Table.Name, tbl.TableAsName)}},
+					Indexes:  []model.CIStr{model.NewCIStr("primary")},
+				})
+			}
+		}
+	case *PhysicalIndexLookUpReader:
+		index := pp.IndexPlans[0].(*PhysicalIndexScan)
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(HintUseIndex),
+			Tables:   []ast.HintTable{{DBName: index.DBName, TableName: getTableName(index.Table.Name, index.TableAsName)}},
+			Indexes:  []model.CIStr{index.Index.Name},
+		})
+		orderHint := HintOrderIndex
+		if !index.KeepOrder {
+			orderHint = HintNoOrderIndex
+		}
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(orderHint),
+			Tables:   []ast.HintTable{{DBName: index.DBName, TableName: getTableName(index.Table.Name, index.TableAsName)}},
+			Indexes:  []model.CIStr{index.Index.Name},
+		})
+	case *PhysicalIndexReader:
+		index := pp.IndexPlans[0].(*PhysicalIndexScan)
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(HintUseIndex),
+			Tables:   []ast.HintTable{{DBName: index.DBName, TableName: getTableName(index.Table.Name, index.TableAsName)}},
+			Indexes:  []model.CIStr{index.Index.Name},
+		})
+		orderHint := HintOrderIndex
+		if !index.KeepOrder {
+			orderHint = HintNoOrderIndex
+		}
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(orderHint),
+			Tables:   []ast.HintTable{{DBName: index.DBName, TableName: getTableName(index.Table.Name, index.TableAsName)}},
+			Indexes:  []model.CIStr{index.Index.Name},
+		})
+	case *PhysicalIndexMergeReader:
+		indexs := make([]model.CIStr, 0, 2)
+		var tableName model.CIStr
+		var tableAsName *model.CIStr
+		for _, partialPlan := range pp.PartialPlans {
+			if index, ok := partialPlan[0].(*PhysicalIndexScan); ok {
+				indexs = append(indexs, index.Index.Name)
+				tableName = index.Table.Name
+				tableAsName = index.TableAsName
+			} else {
+				indexName := model.NewCIStr("PRIMARY")
+				indexs = append(indexs, indexName)
+			}
+		}
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(HintIndexMerge),
+			Tables:   []ast.HintTable{{TableName: getTableName(tableName, tableAsName)}},
+			Indexes:  indexs,
+		})
+	case *PhysicalHashAgg:
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(HintHashAgg),
+		})
+		if storeType == kv.TiKV {
+			res = append(res, &ast.TableOptimizerHint{
+				QBName:   qbName,
+				HintName: model.NewCIStr(HintAggToCop),
+			})
+		}
+	case *PhysicalStreamAgg:
+		res = append(res, &ast.TableOptimizerHint{
+			QBName:   qbName,
+			HintName: model.NewCIStr(HintStreamAgg),
+		})
+		if storeType == kv.TiKV {
+			res = append(res, &ast.TableOptimizerHint{
+				QBName:   qbName,
+				HintName: model.NewCIStr(HintAggToCop),
+			})
+		}
+	case *PhysicalMergeJoin:
+		res = append(res, getJoinHints(p.SCtx(), HintSMJ, p.SelectBlockOffset(), nodeType, pp.children...)...)
+	case *PhysicalHashJoin:
+		// TODO: support the hash_join_build and hash_join_probe hint for auto capture
+		res = append(res, getJoinHints(p.SCtx(), HintHJ, p.SelectBlockOffset(), nodeType, pp.children...)...)
+	case *PhysicalIndexJoin:
+		res = append(res, getJoinHints(p.SCtx(), HintINLJ, p.SelectBlockOffset(), nodeType, pp.children[pp.InnerChildIdx])...)
+	case *PhysicalIndexMergeJoin:
+		res = append(res, getJoinHints(p.SCtx(), HintINLMJ, p.SelectBlockOffset(), nodeType, pp.children[pp.InnerChildIdx])...)
+	case *PhysicalIndexHashJoin:
+		res = append(res, getJoinHints(p.SCtx(), HintINLHJ, p.SelectBlockOffset(), nodeType, pp.children[pp.InnerChildIdx])...)
+	}
+	return res
+}
+
+func removeDuplicatedHints(hints []*ast.TableOptimizerHint) []*ast.TableOptimizerHint {
+	if len(hints) < 2 {
+		return hints
+	}
+	m := make(map[string]struct{}, len(hints))
+	res := make([]*ast.TableOptimizerHint, 0, len(hints))
+	for _, hint := range hints {
+		key := utilhint.RestoreTableOptimizerHint(hint)
+		if _, ok := m[key]; ok {
+			continue
+		}
+		m[key] = struct{}{}
+		res = append(res, hint)
+	}
+	return res
+}
+
+func tableNames2HintTableInfo(ctx sessionctx.Context, hintName string, hintTables []ast.HintTable, p *utilhint.BlockHintProcessor, currentOffset int) []hintTableInfo {
+	if len(hintTables) == 0 {
+		return nil
+	}
+	hintTableInfos := make([]hintTableInfo, 0, len(hintTables))
+	defaultDBName := model.NewCIStr(ctx.GetSessionVars().CurrentDB)
+	isInapplicable := false
+	for _, hintTable := range hintTables {
+		tableInfo := hintTableInfo{
+			dbName:       hintTable.DBName,
+			tblName:      hintTable.TableName,
+			partitions:   hintTable.PartitionList,
+			selectOffset: p.GetHintOffset(hintTable.QBName, currentOffset),
+		}
+		if tableInfo.dbName.L == "" {
+			tableInfo.dbName = defaultDBName
+		}
+		switch hintName {
+		case TiDBMergeJoin, HintSMJ, TiDBIndexNestedLoopJoin, HintINLJ, HintINLHJ, HintINLMJ, TiDBHashJoin, HintHJ, HintLeading:
+			if len(tableInfo.partitions) > 0 {
+				isInapplicable = true
+			}
+		}
+		hintTableInfos = append(hintTableInfos, tableInfo)
+	}
+	if isInapplicable {
+		ctx.GetSessionVars().StmtCtx.AppendWarning(
+			fmt.Errorf("Optimizer Hint %s is inapplicable on specified partitions",
+				restore2JoinHint(hintName, hintTableInfos)))
+		return nil
+	}
+	return hintTableInfos
+}
+
+func restore2TableHint(hintTables ...hintTableInfo) string {
+	buffer := bytes.NewBufferString("")
+	for i, table := range hintTables {
+		buffer.WriteString(table.tblName.L)
+		if len(table.partitions) > 0 {
+			buffer.WriteString(" PARTITION(")
+			for j, partition := range table.partitions {
+				if j > 0 {
+					buffer.WriteString(", ")
+				}
+				buffer.WriteString(partition.L)
+			}
+			buffer.WriteString(")")
+		}
+		if i < len(hintTables)-1 {
+			buffer.WriteString(", ")
+		}
+	}
+	return buffer.String()
+}
+
+func restore2JoinHint(hintType string, hintTables []hintTableInfo) string {
+	if len(hintTables) == 0 {
+		return strings.ToUpper(hintType)
+	}
+	buffer := bytes.NewBufferString("/*+ ")
+	buffer.WriteString(strings.ToUpper(hintType))
+	buffer.WriteString("(")
+	buffer.WriteString(restore2TableHint(hintTables...))
+	buffer.WriteString(") */")
+	return buffer.String()
+}
+
+func restore2IndexHint(hintType string, hintIndex indexHintInfo) string {
+	buffer := bytes.NewBufferString("/*+ ")
+	buffer.WriteString(strings.ToUpper(hintType))
+	buffer.WriteString("(")
+	buffer.WriteString(restore2TableHint(hintTableInfo{
+		dbName:     hintIndex.dbName,
+		tblName:    hintIndex.tblName,
+		partitions: hintIndex.partitions,
+	}))
+	if hintIndex.indexHint != nil && len(hintIndex.indexHint.IndexNames) > 0 {
+		for i, indexName := range hintIndex.indexHint.IndexNames {
+			if i > 0 {
+				buffer.WriteString(",")
+			}
+			buffer.WriteString(" " + indexName.L)
+		}
+	}
+	buffer.WriteString(") */")
+	return buffer.String()
+}
+
+func restore2StorageHint(tiflashTables, tikvTables []hintTableInfo) string {
+	buffer := bytes.NewBufferString("/*+ ")
+	buffer.WriteString(strings.ToUpper(HintReadFromStorage))
+	buffer.WriteString("(")
+	if len(tiflashTables) > 0 {
+		buffer.WriteString("tiflash[")
+		buffer.WriteString(restore2TableHint(tiflashTables...))
+		buffer.WriteString("]")
+		if len(tikvTables) > 0 {
+			buffer.WriteString(", ")
+		}
+	}
+	if len(tikvTables) > 0 {
+		buffer.WriteString("tikv[")
+		buffer.WriteString(restore2TableHint(tikvTables...))
+		buffer.WriteString("]")
+	}
+	buffer.WriteString(") */")
+	return buffer.String()
+}
+
+func extractUnmatchedTables(hintTables []hintTableInfo) []string {
+	var tableNames []string
+	for _, table := range hintTables {
+		if !table.matched {
+			tableNames = append(tableNames, table.tblName.O)
+		}
+	}
+	return tableNames
+}
