+++ pkg/planner/core/planbuilder.go
@@ -15,7 +15,6 @@
 package core
 
 import (
-	"bytes"
 	"context"
 	"encoding/binary"
 	"fmt"
@@ -82,99 +81,6 @@ type visitInfo struct {
 	dynamicWithGrant bool
 }
 
-type indexNestedLoopJoinTables struct {
-	inljTables  []hintTableInfo
-	inlhjTables []hintTableInfo
-	inlmjTables []hintTableInfo
-}
-
-type tableHintInfo struct {
-	indexNestedLoopJoinTables
-	noIndexJoinTables   indexNestedLoopJoinTables
-	sortMergeJoinTables []hintTableInfo
-	broadcastJoinTables []hintTableInfo
-	shuffleJoinTables   []hintTableInfo
-	hashJoinTables      []hintTableInfo
-	noHashJoinTables    []hintTableInfo
-	noMergeJoinTables   []hintTableInfo
-	indexHintList       []indexHintInfo
-	tiflashTables       []hintTableInfo
-	tikvTables          []hintTableInfo
-	aggHints            aggHintInfo
-	indexMergeHintList  []indexHintInfo
-	timeRangeHint       ast.HintTimeRange
-	limitHints          limitHintInfo
-	MergeHints          MergeHintInfo
-	leadingJoinOrder    []hintTableInfo
-	hjBuildTables       []hintTableInfo
-	hjProbeTables       []hintTableInfo
-}
-
-type limitHintInfo struct {
-	preferLimitToCop bool
-}
-
-// MergeHintInfo ...one bool flag for cte
-type MergeHintInfo struct {
-	preferMerge bool
-}
-
-type hintTableInfo struct {
-	dbName       model.CIStr
-	tblName      model.CIStr
-	partitions   []model.CIStr
-	selectOffset int
-	matched      bool
-}
-
-type indexHintInfo struct {
-	dbName     model.CIStr
-	tblName    model.CIStr
-	partitions []model.CIStr
-	indexHint  *ast.IndexHint
-	// Matched indicates whether this index hint
-	// has been successfully applied to a DataSource.
-	// If an indexHintInfo is not matched after building
-	// a Select statement, we will generate a warning for it.
-	matched bool
-}
-
-func (hint *indexHintInfo) match(dbName, tblName model.CIStr) bool {
-	return hint.tblName.L == tblName.L &&
-		(hint.dbName.L == dbName.L ||
-			hint.dbName.L == "*") // for universal bindings, e.g. *.t
-}
-
-func (hint *indexHintInfo) hintTypeString() string {
-	switch hint.indexHint.HintType {
-	case ast.HintUse:
-		return "use_index"
-	case ast.HintIgnore:
-		return "ignore_index"
-	case ast.HintForce:
-		return "force_index"
-	}
-	return ""
-}
-
-// indexString formats the indexHint as dbName.tableName[, indexNames].
-func (hint *indexHintInfo) indexString() string {
-	var indexListString string
-	indexList := make([]string, len(hint.indexHint.IndexNames))
-	for i := range hint.indexHint.IndexNames {
-		indexList[i] = hint.indexHint.IndexNames[i].L
-	}
-	if len(indexList) > 0 {
-		indexListString = fmt.Sprintf(", %s", strings.Join(indexList, ", "))
-	}
-	return fmt.Sprintf("%s.%s%s", hint.dbName, hint.tblName, indexListString)
-}
-
-type aggHintInfo struct {
-	preferAggType  uint
-	preferAggToCop bool
-}
-
 // QueryTimeRange represents a time range specified by TIME_RANGE hint
 type QueryTimeRange struct {
 	From time.Time
@@ -195,234 +101,6 @@ func (tr *QueryTimeRange) MemoryUsage() (sum int64) {
 	}
 
 	return emptyQueryTimeRangeSize
-}
-
-func tableNames2HintTableInfo(ctx sessionctx.Context, hintName string, hintTables []ast.HintTable, p *hint.BlockHintProcessor, currentOffset int) []hintTableInfo {
-	if len(hintTables) == 0 {
-		return nil
-	}
-	hintTableInfos := make([]hintTableInfo, 0, len(hintTables))
-	defaultDBName := model.NewCIStr(ctx.GetSessionVars().CurrentDB)
-	isInapplicable := false
-	for _, hintTable := range hintTables {
-		tableInfo := hintTableInfo{
-			dbName:       hintTable.DBName,
-			tblName:      hintTable.TableName,
-			partitions:   hintTable.PartitionList,
-			selectOffset: p.GetHintOffset(hintTable.QBName, currentOffset),
-		}
-		if tableInfo.dbName.L == "" {
-			tableInfo.dbName = defaultDBName
-		}
-		switch hintName {
-		case TiDBMergeJoin, HintSMJ, TiDBIndexNestedLoopJoin, HintINLJ, HintINLHJ, HintINLMJ, TiDBHashJoin, HintHJ, HintLeading:
-			if len(tableInfo.partitions) > 0 {
-				isInapplicable = true
-			}
-		}
-		hintTableInfos = append(hintTableInfos, tableInfo)
-	}
-	if isInapplicable {
-		ctx.GetSessionVars().StmtCtx.AppendWarning(
-			fmt.Errorf("Optimizer Hint %s is inapplicable on specified partitions",
-				restore2JoinHint(hintName, hintTableInfos)))
-		return nil
-	}
-	return hintTableInfos
-}
-
-func (info *tableHintInfo) ifPreferMergeJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.sortMergeJoinTables)
-}
-
-func (info *tableHintInfo) ifPreferBroadcastJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.broadcastJoinTables)
-}
-
-func (info *tableHintInfo) ifPreferShuffleJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.shuffleJoinTables)
-}
-
-func (info *tableHintInfo) ifPreferHashJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.hashJoinTables)
-}
-
-func (info *tableHintInfo) ifPreferNoHashJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.noHashJoinTables)
-}
-
-func (info *tableHintInfo) ifPreferNoMergeJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.noMergeJoinTables)
-}
-
-func (info *tableHintInfo) ifPreferHJBuild(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.hjBuildTables)
-}
-
-func (info *tableHintInfo) ifPreferHJProbe(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.hjProbeTables)
-}
-
-func (info *tableHintInfo) ifPreferINLJ(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.indexNestedLoopJoinTables.inljTables)
-}
-
-func (info *tableHintInfo) ifPreferINLHJ(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.indexNestedLoopJoinTables.inlhjTables)
-}
-
-func (info *tableHintInfo) ifPreferINLMJ(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.indexNestedLoopJoinTables.inlmjTables)
-}
-
-func (info *tableHintInfo) ifPreferNoIndexJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.noIndexJoinTables.inljTables)
-}
-
-func (info *tableHintInfo) ifPreferNoIndexHashJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.noIndexJoinTables.inlhjTables)
-}
-
-func (info *tableHintInfo) ifPreferNoIndexMergeJoin(tableNames ...*hintTableInfo) bool {
-	return info.matchTableName(tableNames, info.noIndexJoinTables.inlmjTables)
-}
-
-func (info *tableHintInfo) ifPreferTiFlash(tableName *hintTableInfo) *hintTableInfo {
-	if tableName == nil {
-		return nil
-	}
-	for i, tbl := range info.tiflashTables {
-		if tableName.dbName.L == tbl.dbName.L && tableName.tblName.L == tbl.tblName.L && tbl.selectOffset == tableName.selectOffset {
-			info.tiflashTables[i].matched = true
-			return &tbl
-		}
-	}
-	return nil
-}
-
-func (info *tableHintInfo) ifPreferTiKV(tableName *hintTableInfo) *hintTableInfo {
-	if tableName == nil {
-		return nil
-	}
-	for i, tbl := range info.tikvTables {
-		if tableName.dbName.L == tbl.dbName.L && tableName.tblName.L == tbl.tblName.L && tbl.selectOffset == tableName.selectOffset {
-			info.tikvTables[i].matched = true
-			return &tbl
-		}
-	}
-	return nil
-}
-
-// matchTableName checks whether the hint hit the need.
-// Only need either side matches one on the list.
-// Even though you can put 2 tables on the list,
-// it doesn't mean optimizer will reorder to make them
-// join directly.
-// Which it joins on with depend on sequence of traverse
-// and without reorder, user might adjust themselves.
-// This is similar to MySQL hints.
-func (*tableHintInfo) matchTableName(tables []*hintTableInfo, hintTables []hintTableInfo) bool {
-	hintMatched := false
-	for _, table := range tables {
-		for i, curEntry := range hintTables {
-			if table == nil {
-				continue
-			}
-			if (curEntry.dbName.L == table.dbName.L || curEntry.dbName.L == "*") &&
-				curEntry.tblName.L == table.tblName.L &&
-				table.selectOffset == curEntry.selectOffset {
-				hintTables[i].matched = true
-				hintMatched = true
-				break
-			}
-		}
-	}
-	return hintMatched
-}
-
-func restore2TableHint(hintTables ...hintTableInfo) string {
-	buffer := bytes.NewBufferString("")
-	for i, table := range hintTables {
-		buffer.WriteString(table.tblName.L)
-		if len(table.partitions) > 0 {
-			buffer.WriteString(" PARTITION(")
-			for j, partition := range table.partitions {
-				if j > 0 {
-					buffer.WriteString(", ")
-				}
-				buffer.WriteString(partition.L)
-			}
-			buffer.WriteString(")")
-		}
-		if i < len(hintTables)-1 {
-			buffer.WriteString(", ")
-		}
-	}
-	return buffer.String()
-}
-
-func restore2JoinHint(hintType string, hintTables []hintTableInfo) string {
-	if len(hintTables) == 0 {
-		return strings.ToUpper(hintType)
-	}
-	buffer := bytes.NewBufferString("/*+ ")
-	buffer.WriteString(strings.ToUpper(hintType))
-	buffer.WriteString("(")
-	buffer.WriteString(restore2TableHint(hintTables...))
-	buffer.WriteString(") */")
-	return buffer.String()
-}
-
-func restore2IndexHint(hintType string, hintIndex indexHintInfo) string {
-	buffer := bytes.NewBufferString("/*+ ")
-	buffer.WriteString(strings.ToUpper(hintType))
-	buffer.WriteString("(")
-	buffer.WriteString(restore2TableHint(hintTableInfo{
-		dbName:     hintIndex.dbName,
-		tblName:    hintIndex.tblName,
-		partitions: hintIndex.partitions,
-	}))
-	if hintIndex.indexHint != nil && len(hintIndex.indexHint.IndexNames) > 0 {
-		for i, indexName := range hintIndex.indexHint.IndexNames {
-			if i > 0 {
-				buffer.WriteString(",")
-			}
-			buffer.WriteString(" " + indexName.L)
-		}
-	}
-	buffer.WriteString(") */")
-	return buffer.String()
-}
-
-func restore2StorageHint(tiflashTables, tikvTables []hintTableInfo) string {
-	buffer := bytes.NewBufferString("/*+ ")
-	buffer.WriteString(strings.ToUpper(HintReadFromStorage))
-	buffer.WriteString("(")
-	if len(tiflashTables) > 0 {
-		buffer.WriteString("tiflash[")
-		buffer.WriteString(restore2TableHint(tiflashTables...))
-		buffer.WriteString("]")
-		if len(tikvTables) > 0 {
-			buffer.WriteString(", ")
-		}
-	}
-	if len(tikvTables) > 0 {
-		buffer.WriteString("tikv[")
-		buffer.WriteString(restore2TableHint(tikvTables...))
-		buffer.WriteString("]")
-	}
-	buffer.WriteString(") */")
-	return buffer.String()
-}
-
-func extractUnmatchedTables(hintTables []hintTableInfo) []string {
-	var tableNames []string
-	for _, table := range hintTables {
-		if !table.matched {
-			tableNames = append(tableNames, table.tblName.O)
-		}
-	}
-	return tableNames
 }
 
 // clauseCode indicates in which clause the column is currently.
@@ -515,14 +193,6 @@ const (
 	handlingCompareSubquery
 	handlingInSubquery
 	handlingScalarSubquery
-)
-
-// Hint flags listed here are used by PlanBuilder.subQueryHintFlags.
-const (
-	// HintFlagSemiJoinRewrite corresponds to HintSemiJoinRewrite.
-	HintFlagSemiJoinRewrite uint64 = 1 << iota
-	// HintFlagNoDecorrelate corresponds to HintNoDecorrelate.
-	HintFlagNoDecorrelate
 )
 
 // PlanBuilder builds Plan from an ast.Node.
